refactor to follow naming guideline
