+++ pages/_app.js
@@ -1,9 +1,7 @@
 import "../styles/globals.css";
 
-function App({ Component, pageProps }) {
+export default function app({ Component, pageProps }) {
   const getLayout = Component.getLayout || ((page) => page);
 
   return getLayout(<Component {...pageProps} />);
 }
-
-export default App;
